requirements.txt
PyYAML
confluent-kafka
psycopg2-binary
aiobotocore==2.4.2
boto3==1.24.59
botocore==1.27.59